Stage

+++ src/app/(navigation)/(right-search)/notification/schema.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod'
 
 export const markAsReadSchema = z.object({
-  ids: z.array(z.coerce.number().positive()).min(1),
+  ids: z.array(z.coerce.number().int().positive()).min(1),
 })
 
 export const deleteNotificationsSchema = z.object({
-  ids: z.array(z.coerce.number().positive()).min(1),
+  ids: z.array(z.coerce.number().int().positive()).min(1),
 })
