+++ src/app/manga/[id]/[source]/schema.ts
@@ -3,6 +3,6 @@ import { z } from 'zod/v4'
 import { SourceParam } from '@/utils/param'
 
 export const mangaSchema = z.object({
-  id: z.coerce.number().positive(),
+  id: z.coerce.number().int().positive(),
   source: z.enum(SourceParam),
 })
